Update to calling CrysFML routines from C.
In order to be compatible with Igor Pro 6.2 on MacOSX10.6, everything has to be compiled with the -m32 option, as Igor is not yet 64-bit compatible, and all compilers as gfortran and gcc otherwise create by default 64-bit code. Also, Scripts/MacOS/make_crysfml , needs to be changed: OPT1="-m32 -c -O" etc. (4 lines).

gfortran creates shared a library libF4C.so which has to be placed in /usr/local/lib. Then Igor can load the XOP calling that library. However, the library needs to be an original file, not an alias!


git-svn-id: https://forge.ill.fr/svn/crysfml@420 06508225-4b59-0410-a2a9-fdab55894d57