CFML_HDF5 for the old version of CrysFML.
In set_compiler_flags I have commented the command
set(CMAKE_Fortran_FLAGS ""). In this way, the compilation
follows the default flags of cmake. Some of these flags are
/libs:dll /threads. This means that a dynamic runtime library
is used. In this way we avoid conflicts when compiling external
programs that make use of CFML_HDF5. If it is preferred to keep
this command, then, when compiling an external program that makes
use of CFML_HDF5, the following flags are required for linking:
 /link /nodefaultlib:library, where library is the library
that enters in conflict.